新增商户游戏表

+++ models/merchantPlatform.go
@@ -7,8 +7,8 @@ const TableNameMerchantPlatform = "merchant_platform"
 // MerchantPlatform undefined
 type MerchantPlatform struct {
 	MerchantPlatformId int64           `json:"merchant_platform_id" gorm:"merchant_platform_id,primaryKey"`
-	MerchantId         int             `json:"merchant_id" gorm:"merchant_id"` // 商户id
-	PlatformId         int             `json:"platform_id" gorm:"platform_id"` // 平台id
+	MerchantId         int64           `json:"merchant_id" gorm:"merchant_id"` // 商户id
+	PlatformId         int64           `json:"platform_id" gorm:"platform_id"` // 平台id
 	Rate               decimal.Decimal `json:"rate" gorm:"rate"`               // 费率
 	Status             int             `json:"status" gorm:"status"`           // 状态 1 启用 2禁用
 	CreatedAt          int64           `json:"created_at" gorm:"created_at"`
